01_etl/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
backoff==1.11.1
    # via -r requirements.in
certifi==2021.10.8
    # via elasticsearch
deprecated==1.2.13
    # via redis
elasticsearch==7.17.0
    # via -r requirements.in
packaging==21.3
    # via redis
psycopg2-binary==2.9.3
    # via -r requirements.in
pyparsing==3.0.7
    # via packaging
redis==4.1.4
    # via -r requirements.in
urllib3==1.26.8
    # via elasticsearch
wrapt==1.14.0
    # via deprecated
